Allow ZipLocator to return slice ownership back in error
When parsing a zip from a vec, it would be nice to return the vec to the caller
on error so that they may continue processing the vec (ie: try and detect if
the data represents a zip and if not, do something else).

This commit also removes the `ReaderError` and instead uses a tuple. I'm not
sure if this is the right approach, but at least this is what reddit thinks :p

https://www.reddit.com/r/rust/comments/n4puxb/function_that_takes_ownership_and_returns_result/