Implement Rust builder API for ZIP archive writer
Replace ZipEntryOptions with method-chaining builders for a slightly
more type safe API. Previously both files and directories could receive
the same zip options, which could result in one specifying a compression
method for directories -- and while the code would nullify it before it
actually was written, having two separate builders makes the creation
more airtight.

This is a breaking change as ZipEntryOptions is no longer public.